schema-to-database/python-server/requirements.txt
JSONSchema2DB==1.0.1
python-socketio==4.4.0  